+++ api/src/main/java/ru/gb/market/api/dto/ProductDto.java
@@ -1,13 +1,15 @@
 package ru.gb.market.api.dto;
 
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private Long id;
 
     private String title;
 
-    private int price;
+    private BigDecimal price;
 
     private String categoryTitle;
 
@@ -24,10 +26,10 @@ public class ProductDto {
     public void setTitle(String title) {
         this.title = title;
     }
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
     public String getCategoryTitle() {
@@ -39,7 +41,7 @@ public class ProductDto {
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, int price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
