+++ product_service/src/main/resources/db/changelog/ddl/mainDB.sql
@@ -15,7 +15,7 @@ create table products
     id          bigserial primary key,
     category_id bigint references categories (id),
     title       varchar(255),
-    price       int,
+    price       numeric (8, 2),
     created_at  timestamp default current_timestamp,
     updated_at  timestamp default current_timestamp
